requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
firebase-admin==6.5.0
anthropic==0.29.0
python-dotenv==1.0.1
gunicorn==22.0.0